+++ CHANGELOG.md
@@ -8,7 +8,6 @@ All notable changes to this plugin will be documented in this file.
 ### Changed
 * Doctor images now use aspect-ratio instead of fixed height. This allows for more flexibility in image sizes.
 ### Fixed
-
 ## [2.0.0] - 2023-06-16
 #### Glacial theme v3.0.0 or later required
 ### Added
@@ -22,3 +21,4 @@ All notable changes to this plugin will be documented in this file.
 ### Changed
 * Both CPT (Doctors and Locations) are now created in ACF settings. This makes it much easier to change things like slugs and labels.
 * Links to doctor services and location now displayed as buttons.
+* CSS: All doctors and related posts containers converted from flex-box to grid.